feat(mlt): introduce watchable steps
Only watchable steps can be used in --watch mode. This change is connected with a deeper refactor of
the whole flow of the package: choosing steps for the run is moved to its own file, additionally the
public API changed.

BREAKING CHANGE: Changes in public API
